add folders for screenshots and  logs
