+++ framework_tests/DevConfig.json
@@ -1,11 +1,11 @@
 ﻿{
-  "Number Of Instances": 2,
+  "NumberOfInstances": 2,
   "SoftWare": "Free: Debian, CentOS, CoreOS, Ubuntu",
-  "VM Class": "Regular",
-  "Instance Type": "n1-standard-2",
-  "Number Of GPUs": 4,
-  "GPU Type": "NVIDIA Tesla P4",
-  "Local SSD": " 2x375 Gb",
-  "Datacenter Location": "London (europe-west2)",
-  "Commited Usage": "1 Year"
+  "VMClass": "Regular",
+  "InstanceType": "n1-standard-2",
+  "NumberOfGPUs": 4,
+  "GPUType": "NVIDIA Tesla P4",
+  "LocalSSD": "2x375 Gb",
+  "DatacenterLocation": "London (europe-west2)",
+  "CommitedUsage": "1 Year"
 }
