+++ framework_tests/Models/ComputerEngineForm.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Runtime.Serialization;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 
@@ -7,19 +6,12 @@ namespace framework_tests
 {
     public class ComputerEngineForm
     {
-        [JsonProperty(PropertyName = "NumberOfInstances")]
         public int NumberOfInstances { get; set; }
         public string SoftWare { get; set; }
         public string VMClass { get; set; }
         public string InstanceType { get; set; }
         public int NumberOfGPUs { get; set; }
         public string GPUType { get; set; }
-
-        public static explicit operator ComputerEngineForm(Task<ComputerEngineForm> v)
-        {
-            throw new NotImplementedException();
-        }
-
         public string LocalSSD { get; set; }
         public string DatacenterLocation { get; set; }
         public string CommitedUsage { get; set; }
