+++ framework_tests/Pages/GCPlatformPricingCalculator.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Threading;
 using NLog;
 using OpenQA.Selenium;
