+++ framework_tests/QAConfig.json
@@ -5,7 +5,7 @@
   "InstanceType": "n1-standard-8",
   "NumberOfGPUs": 1,
   "GPUType": "NVIDIA Tesla V100",
-  "LocalSSD": " 2x375 Gb",
+  "LocalSSD": "2x375 Gb",
   "DatacenterLocation": "Frankfurt (europe-west3)",
   "CommitedUsage": "1 Year"
 }
